Reorganização dos passos em classes distintas.
