+++ impl/parser/jbehave/src/main/java/br/gov/frameworkdemoiselle/behave/parser/jbehave/MouseControlSteps.java
@@ -1,0 +1,81 @@
+/*
+ * Demoiselle Framework
+ * Copyright (C) 2013 SERPRO
+ * ----------------------------------------------------------------------------
+ * This file is part of Demoiselle Framework.
+ * 
+ * Demoiselle Framework is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public License version 3
+ * as published by the Free Software Foundation.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public License version 3
+ * along with this program; if not,  see <http://www.gnu.org/licenses/>
+ * or write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA  02110-1301, USA.
+ * ----------------------------------------------------------------------------
+ * Este arquivo é parte do Framework Demoiselle.
+ * 
+ * O Framework Demoiselle é um software livre; você pode redistribuí-lo e/ou
+ * modificá-lo dentro dos termos da GNU LGPL versão 3 como publicada pela Fundação
+ * do Software Livre (FSF).
+ * 
+ * Este programa é distribuído na esperança que possa ser útil, mas SEM NENHUMA
+ * GARANTIA; sem uma garantia implícita de ADEQUAÇÃO a qualquer MERCADO ou
+ * APLICAÇÃO EM PARTICULAR. Veja a Licença Pública Geral GNU/LGPL em português
+ * para maiores detalhes.
+ * 
+ * Você deve ter recebido uma cópia da GNU LGPL versão 3, sob o título
+ * "LICENCA.txt", junto com esse programa. Se não, acesse <http://www.gnu.org/licenses/>
+ * ou escreva para a Fundação do Software Livre (FSF) Inc.,
+ * 51 Franklin St, Fifth Floor, Boston, MA 02111-1301, USA.
+ */
+
+package br.gov.frameworkdemoiselle.behave.parser.jbehave;
+
+
+import org.jbehave.core.annotations.When;
+
+import br.gov.frameworkdemoiselle.behave.dataprovider.DataProvider;
+import br.gov.frameworkdemoiselle.behave.dataprovider.DatasetProvider;
+import br.gov.frameworkdemoiselle.behave.runner.Runner;
+import br.gov.frameworkdemoiselle.behave.runner.ui.Button;
+import br.gov.frameworkdemoiselle.behave.runner.ui.Element;
+import br.gov.frameworkdemoiselle.behave.runner.ui.Link;
+import br.gov.frameworkdemoiselle.behave.runner.ui.Menu;
+import br.gov.frameworkdemoiselle.behave.runner.ui.MenuItem;
+
+public class MouseControlSteps extends CommonSteps{
+
+	
+	protected Runner runner;
+	protected DataProvider dataProvider;
+	protected DatasetProvider datasetProvider;
+	protected static String currentPageName;
+		
+	public MouseControlSteps() {
+		super();
+		this.runner = super.runner;
+		this.dataProvider = super.dataProvider;
+		this.datasetProvider = super.datasetProvider;
+		
+	}
+	
+	@When("movo o mouse sobre \"$element\"")
+	public void moverMouse(String elementName) {
+		Element element = runner.getElement(currentPageName, elementName);
+		if (element instanceof Menu) {
+			((Menu) element).mouseOver();
+		} else if (element instanceof MenuItem) {
+			((MenuItem) element).mouseOver();
+		} else if (element instanceof Button) {
+			((Button) element).mouseOver();
+		} else if (element instanceof Link) {
+			((Link) element).mouseOver();
+		}
+	}
+}
